feat(diagnostics): split combined API route, decouple panels into independent fetches
The old GET /api/admin/diagnostics did one big fetch backing all 5
Diagnostics panels off a single loading/error state -- retrying one
sync issue or resuming one suspended meeting reloaded all five.

Split into 5 routes (system-status, meeting-counts, sync-issues,
conflicts, suspended), each with its own scoped Prisma query.
DiagnosticsTab.tsx is now a thin shell composing 5 independent card
components, each owning its own fetch/loading/error state and showing
its own TopLoadingBar (reused from the calendar views) during load.

Each card also carries its own out-of-order-response guard, same
pattern the old combined fetch used, since actions can trigger a
second load() before a first one resolves.
