+++ frontend/app/components/admin/DiagnosticsTab.tsx
@@ -1,13 +1,12 @@
 "use client";
 
-import React, { useEffect, useRef, useState } from "react";
+import React from "react";
 import type { Role } from "@prisma/client";
-import StatCounter from "../atoms/StatCounter";
-import Card from "./Card";
-import ConflictList, { ConflictListRow } from "./ConflictList";
-import ResumeMeetingModal from "../meeting-form/ResumeMeetingModal";
-import { formatSuspensionStatusText } from "../../../util/suspensionText";
-import { retryMeetingSync } from "../../../util/syncMeeting";
+import SystemStatusCard from "./SystemStatusCard";
+import MeetingCountsCard from "./MeetingCountsCard";
+import SyncIssuesCard from "./SyncIssuesCard";
+import ConflictsCard from "./ConflictsCard";
+import SuspendedCard from "./SuspendedCard";
 import styles from "../../../styles/components/admin/DiagnosticsTab.module.scss";
 
 interface DiagnosticsTabProps {
@@ -15,347 +14,17 @@ interface DiagnosticsTabProps {
   role: Role;
 }
 
-interface DiagnosticsData {
-  database: { ok: boolean; latencyMs: number | null };
-  googleCalendar: { categories: Record<string, boolean> };
-  zoom: {
-    reachable: boolean;
-    roomCalendars: Record<string, boolean>;
-    hostPool: Record<string, { ok: boolean; licensed: boolean | null }>;
-  };
-  session: { email: string | null; role: Role | null };
-  meetingCounts: {
-    total: number;
-    active: number;
-    suspended: number;
-    suspendedOrPending: number;
-    byCategory: Record<string, number>;
-    recurring: number;
-    oneTime: number;
-    gcalSyncErrors: number;
-    zoomSyncErrors: number;
-    pendingZoomSync: number;
-  };
-  conflicts: ConflictListRow[];
-  syncIssues: {
-    mid: string;
-    title: string;
-    group: string;
-    room: string;
-    modeType: string;
-    calType: string[];
-    issues: string[];
-    updatedAt: string | null;
-  }[];
-  suspendedMeetings: {
-    mid: string;
-    title: string;
-    group: string;
-    room: string;
-    modeType: string;
-    calType: string[];
-    updatedAt: string | null;
-    resumesAt: string | null;
-    suspendedSince: string | null;
-    suspensionActive: boolean;
-  }[];
-}
-
-const roleLabel: Record<string, string> = {
-  SUPER_ADMIN: "Super Admin",
-  ADMIN: "Admin",
-};
-
-const DiagnosticsTab: React.FC<DiagnosticsTabProps> = ({ email, role }) => {
-  const [data, setData] = useState<DiagnosticsData | null>(null);
-  const [error, setError] = useState<string | null>(null);
-  // Guards against out-of-order resolution: loadDiagnostics is called both on mount and from
-  // ConflictList's onMeetingUpdated, so a slow initial fetch could otherwise resolve after (and
-  // overwrite) a faster post-edit refresh.
-  const latestRequestId = useRef(0);
-  // Which sync-issue row's "Retry sync" is in flight, so only that row shows "Retrying…"
-  // instead of every row disabling at once.
-  const [retryingMid, setRetryingMid] = useState<string | null>(null);
-  // Which suspended-meeting row's "Resume" is in flight, same pattern as retryingMid above.
-  const [resumingMid, setResumingMid] = useState<string | null>(null);
-  // The suspended row currently showing its ResumeMeetingModal (Immediately vs. On a date),
-  // if any -- { mid, title } rather than just the id since the modal needs the title too.
-  const [resumeModalMeeting, setResumeModalMeeting] = useState<{ mid: string; title: string; suspendedSince: string | null; suspensionActive: boolean } | null>(null);
-
-  const loadDiagnostics = async () => {
-    const requestId = ++latestRequestId.current;
-    try {
-      const response = await fetch("/api/admin/diagnostics");
-      if (!response.ok) throw new Error(`HTTP error! status: ${response.status}`);
-      const json: DiagnosticsData = await response.json();
-      if (requestId === latestRequestId.current) setData(json);
-    } catch (err) {
-      console.error("Error fetching diagnostics:", err);
-      if (requestId === latestRequestId.current) setError("Failed to load diagnostics.");
-    }
-  };
-
-  useEffect(() => {
-    // loadDiagnostics is also called from ConflictList's onMeetingUpdated, so it can't be
-    // defined inline inside this effect the way the lint rule expects.
-    // eslint-disable-next-line react-hooks/set-state-in-effect
-    loadDiagnostics();
-  }, []);
-
-  // Same endpoint/shape as ViewMeeting.tsx's "Retry sync" button -- reused here so a Sync
-  // Issues row can be retried without navigating to the meeting's own detail view first.
-  const retrySync = async (mid: string) => {
-    setRetryingMid(mid);
-    try {
-      await retryMeetingSync(mid);
-      await loadDiagnostics();
-    } catch (err) {
-      console.error("Error retrying sync:", err);
-    } finally {
-      // Only clear if this is still the row that started this retry -- a second row's
-      // retry starting before this one resolves must not un-disable/mislabel it early.
-      setRetryingMid((current) => (current === mid ? null : current));
-    }
-  };
-
-  // Same endpoint ViewMeeting.tsx's kebab-menu "Reactivate" calls -- reused here so a suspended
-  // meeting can be resumed straight from the Diagnostics panel too. `on` is omitted for an
-  // immediate resume, or an ISO date string to schedule the resume instead (see
-  // ResumeMeetingModal / the resume route's `on` branch).
-  const resumeMeeting = async (mid: string, on: string | null) => {
-    setResumingMid(mid);
-    try {
-      const response = await fetch("/api/update/meeting/resume", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(on ? { mid, on } : { mid }),
-      });
-      if (!response.ok) {
-        // Previously silent on failure (console.error only) -- a rejected date (e.g. "on" before
-        // the suspension's own start) would just close the modal with no visible feedback at all.
-        const body = await response.json().catch(() => null);
-        throw new Error(body?.error || `HTTP error! status: ${response.status}`);
-      }
-      await loadDiagnostics();
-    } catch (err) {
-      console.error("Error resuming meeting:", err);
-      alert(`Error: could not resume the meeting${err instanceof Error ? ` (${err.message})` : ""}`);
-    } finally {
-      setResumingMid((current) => (current === mid ? null : current));
-    }
-  };
-
-  if (error) return <Card>{error}</Card>;
-  if (!data) return <Card>Loading diagnostics…</Card>;
-
-  const gcalEntries = Object.entries(data.googleCalendar.categories);
-  const gcalReachableCount = gcalEntries.filter(([, ok]) => ok).length;
-
-  const roomCalendarEntries = Object.entries(data.zoom.roomCalendars);
-  const roomCalendarOkCount = roomCalendarEntries.filter(([, ok]) => ok).length;
-
-  const hostPoolEntries = Object.entries(data.zoom.hostPool);
-  const hostPoolReachableCount = hostPoolEntries.filter(([, s]) => s.ok).length;
-  // Confirmed Licensed only (licensed === true) -- licensed === null means the check couldn't
-  // determine license status (e.g. Zoom unreachable), and counting that alongside a confirmed
-  // Licensed host would overstate this line, which reads as a positive claim ("N licensed").
-  const hostPoolLicensedCount = hostPoolEntries.filter(([, s]) => s.licensed === true).length;
-
-  return (
-    <div className={styles.container}>
-      <Card>
-        <div className={styles.sectionLabel}>SYSTEM STATUS</div>
-
-        <div className={styles.statusRow}>
-          <span className={`${styles.dot} ${data.database.ok ? styles.dotOk : styles.dotDown}`} />
-          <span className={styles.statusLabel}>Database</span>
-          <span className={styles.statusDetail}>
-            {data.database.ok ? `Connected · ${data.database.latencyMs}ms` : "Unreachable"}
-          </span>
-        </div>
-
-        <div className={styles.statusRow}>
-          <span className={`${styles.dot} ${gcalReachableCount === gcalEntries.length ? styles.dotOk : styles.dotDown}`} />
-          <span className={styles.statusLabel}>Google Calendar</span>
-          <span className={styles.statusDetail}>
-            {gcalReachableCount}/{gcalEntries.length} calendars reachable
-          </span>
-        </div>
-        <div className={styles.gcalSubRow}>
-          {gcalEntries.map(([cat, ok]) => (
-            <span key={cat} className={ok ? styles.gcalOk : styles.gcalDown}>
-              {cat}: {ok ? "✓" : "✕ unreachable"}
-            </span>
-          ))}
-        </div>
-
-        <div className={styles.statusRow}>
-          <span className={`${styles.dot} ${data.zoom.reachable ? styles.dotOk : styles.dotDown}`} />
-          <span className={styles.statusLabel}>Zoom</span>
-          <span className={styles.statusDetail}>
-            {data.zoom.reachable ? "Account reachable" : "Account unreachable"}
-            {" · "}{roomCalendarOkCount}/{roomCalendarEntries.length} room calendars reachable
-            {" · "}{hostPoolReachableCount}/{hostPoolEntries.length} pooled hosts reachable
-            {" · "}{hostPoolLicensedCount}/{hostPoolEntries.length} licensed
-          </span>
-        </div>
-        <div className={styles.gcalSubRow}>
-          {roomCalendarEntries.map(([room, ok]) => (
-            <span key={room} className={ok ? styles.gcalOk : styles.gcalDown}>
-              {room.replace(/ - Zoom$/, "")}: {ok ? "✓" : "✕ unreachable"}
-            </span>
-          ))}
-        </div>
-        <div className={styles.gcalSubRow}>
-          {hostPoolEntries.map(([host, s]) => {
-            let detail = s.ok ? "✓" : "✕ unreachable";
-            if (s.ok && s.licensed === false) detail += " (Basic — 40min cap)";
-            return (
-              <span key={host} className={s.ok && s.licensed !== false ? styles.gcalOk : styles.gcalDown}>
-                {host}: {detail}
-              </span>
-            );
-          })}
-        </div>
-
-        <div className={styles.statusRow}>
-          <span className={styles.dot} style={{ backgroundColor: "#4CAF50" }} />
-          <span className={styles.statusLabel}>Session</span>
-          <span className={styles.statusDetail}>
-            Active · {email} — Role: {roleLabel[role] ?? role}
-          </span>
-        </div>
-      </Card>
-
-      <Card>
-        <div className={styles.sectionLabel}>MEETING COUNTS</div>
-        <div className={styles.countsRow}>
-          <StatCounter value={data.meetingCounts.total} label="Total" />
-          <StatCounter value={data.meetingCounts.active} label="Active" />
-          <StatCounter value={data.meetingCounts.suspended} label="Suspended" variant="warning" />
-        </div>
-        <div className={styles.countsSecondaryRow}>
-          <span>
-            {Object.entries(data.meetingCounts.byCategory)
-              .map(([cat, count]) => `${cat}: ${count}`)
-              .join(" · ")}
-          </span>
-          <span>
-            Recurring: {data.meetingCounts.recurring} · One-time: {data.meetingCounts.oneTime}
-          </span>
-        </div>
-        {(data.meetingCounts.gcalSyncErrors > 0 || data.meetingCounts.zoomSyncErrors > 0) && (
-          <div className={styles.countsSecondaryRow}>
-            <span className={styles.gcalDown}>
-              ⚠ Sync errors — Google Calendar: {data.meetingCounts.gcalSyncErrors} · Zoom: {data.meetingCounts.zoomSyncErrors}
-            </span>
-          </div>
-        )}
-        {data.meetingCounts.pendingZoomSync > 0 && (
-          <div className={styles.countsSecondaryRow}>
-            <span className={styles.gcalDown}>
-              ⏳ Waiting on a Zoom host: {data.meetingCounts.pendingZoomSync} — calendars
-              won&apos;t publish until a host becomes available (see Sync Issues below to retry)
-            </span>
-          </div>
-        )}
-      </Card>
-
-      <Card data-testid="diagnostics-sync-issues-panel">
-        <div className={styles.panelHeader}>⚠ Sync Issues ({data.syncIssues.length})</div>
-        <div className={styles.panelSubhead}>
-          Meetings that failed to sync to Zoom or Google Calendar, or are waiting on a Zoom host to
-          become available. Retry here, or open the meeting to edit it if retrying doesn&apos;t resolve it.
-        </div>
-        {data.syncIssues.length === 0 ? (
-          <div className={styles.emptyState}>No sync issues.</div>
-        ) : (
-          data.syncIssues.map((meeting) => (
-            <div key={meeting.mid} className={styles.meetingRow}>
-              <div className={styles.syncIssueRow}>
-                <div>
-                  <span className={styles.meetingTitle}>{meeting.title}</span>{" "}
-                  <span className={styles.meetingTags}>({meeting.group})</span>
-                  <div className={styles.meetingMeta}>
-                    {meeting.room} · {meeting.modeType} · {meeting.calType.join(", ")}
-                  </div>
-                  {meeting.issues.map((issue) => (
-                    <div key={issue} className={styles.issueLine}>{issue}</div>
-                  ))}
-                </div>
-                <button
-                  type="button"
-                  className={styles.retryButton}
-                  onClick={() => retrySync(meeting.mid)}
-                  disabled={retryingMid === meeting.mid}
-                >
-                  {retryingMid === meeting.mid ? "Retrying…" : "Retry sync"}
-                </button>
-              </div>
-            </div>
-          ))
-        )}
-      </Card>
-
-      <Card data-testid="diagnostics-conflicts-panel">
-        <div className={styles.panelHeader}>⚠ Conflicts ({data.conflicts.length})</div>
-        <div className={styles.panelSubhead}>
-          These meetings share a room, Zoom room, or Zoom host at overlapping times. Review and edit one to resolve.
-        </div>
-        <ConflictList conflicts={data.conflicts} onMeetingUpdated={loadDiagnostics} />
-      </Card>
-
-      <Card data-testid="diagnostics-suspended-panel">
-        <div className={styles.panelHeader}>⏸ Suspended ({data.meetingCounts.suspendedOrPending})</div>
-        <div className={styles.panelSubhead}>
-          Meetings currently suspended, or with a suspension scheduled for a future date. Active
-          ones are hidden from the live calendar and Google Calendar; scheduled ones still show
-          normally until their start date arrives. All remain in the system and can be reactivated
-          (or have a scheduled suspension cancelled) here or from the meeting itself.
-        </div>
-        {data.suspendedMeetings.length === 0 ? (
-          <div className={styles.emptyState}>No suspended meetings.</div>
-        ) : (
-          data.suspendedMeetings.map((meeting) => (
-            <div key={meeting.mid} className={styles.meetingRow}>
-              <div className={styles.syncIssueRow}>
-                <div>
-                  <span className={styles.meetingTitle}>{meeting.title}</span>{" "}
-                  <span className={styles.meetingTags}>({meeting.calType.join(", ")})</span>
-                  <div className={styles.meetingMeta}>
-                    {meeting.room} · {meeting.modeType} · {meeting.calType.join(", ")}
-                  </div>
-                  <div className={styles.issueLine}>
-                    {formatSuspensionStatusText(meeting.suspendedSince, meeting.resumesAt, meeting.suspensionActive)}
-                  </div>
-                </div>
-                <button
-                  type="button"
-                  className={styles.retryButton}
-                  onClick={() => setResumeModalMeeting({ mid: meeting.mid, title: meeting.title, suspendedSince: meeting.suspendedSince, suspensionActive: meeting.suspensionActive })}
-                  disabled={resumingMid === meeting.mid}
-                >
-                  {resumingMid === meeting.mid ? "Resuming…" : meeting.suspensionActive ? "Resume" : "Cancel"}
-                </button>
-              </div>
-            </div>
-          ))
-        )}
-      </Card>
-
-      <ResumeMeetingModal
-        isOpen={resumeModalMeeting !== null}
-        title={resumeModalMeeting?.title ?? ""}
-        suspendedSince={resumeModalMeeting?.suspendedSince}
-        isActive={resumeModalMeeting?.suspensionActive ?? true}
-        onCancel={() => setResumeModalMeeting(null)}
-        onConfirm={(on) => {
-          if (resumeModalMeeting) resumeMeeting(resumeModalMeeting.mid, on);
-          setResumeModalMeeting(null);
-        }}
-      />
-    </div>
-  );
-};
+// Each panel below fetches and refreshes independently (its own endpoint under
+// /api/admin/diagnostics/*, own loading state) rather than sharing one combined fetch --
+// e.g. retrying a sync issue only reloads SyncIssuesCard, not the other four.
+const DiagnosticsTab: React.FC<DiagnosticsTabProps> = ({ email, role }) => (
+  <div className={styles.container}>
+    <SystemStatusCard email={email} role={role} />
+    <MeetingCountsCard />
+    <SyncIssuesCard />
+    <ConflictsCard />
+    <SuspendedCard />
+  </div>
+);
 
 export default DiagnosticsTab;
